% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedMethodsComparison.R
\name{alignmentPASTE2}
\alias{alignmentPASTE2}
\title{alignmentPASTE2(object)}
\usage{
alignmentPASTE2(
  object,
  patientType = c("unique", "multiple"),
  samples = c("19", "43", "47", "45")
)
}
\arguments{
\item{object}{first object with the slices/samples merged together}

\item{patientType}{Type of patient data. Options are 'unique' or 'multiple'.}

\item{samples}{Vector of sample identifiers. Default is c("19","43","47","45"). It is needed for multiple patients.}
}
\description{
alignmentPASTE2(object)
}
