% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{as_AssayObject}
\alias{as_AssayObject}
\title{as_AssayObject(object)}
\usage{
as_AssayObject(object)
}
\arguments{
\item{object}{An RCTD object.}
}
\value{
A Seurat Assay object containing the deconvolution results.
}
\description{
Converts an RCTD object to a Seurat Assay object.
}
