% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{calcNewCoord}
\alias{calcNewCoord}
\title{calcNewCoord(coord, sol, xmax, ymax)}
\usage{
calcNewCoord(coord, sol, xmax, ymax)
}
\arguments{
\item{coord}{original coordinates}

\item{sol}{calculated transformation parameters}

\item{xmax}{image dimensions}

\item{ymax}{image dimensions}
}
\value{
new coordinates matrix
}
\description{
Given the original coordinates (coord), calculated transformation parameters (sol),
and image dimensions (xmax, ymax), this function calculates the new transformed coordinates.
}
