% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{calcParameters}
\alias{calcParameters}
\title{calcParameters(solution, xmax, ymax)}
\usage{
calcParameters(solution, xmax, ymax)
}
\arguments{
\item{solution}{}

\item{xmax}{}

\item{ymax}{}
}
\value{
transformation parameters
}
\description{
Given the solution and the image dimensions (xmax, ymax),
this function calculates all necessary parameters for use in semla functions.
}
