% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{calcScal}
\alias{calcScal}
\title{calcScal(coord1, coord2)}
\usage{
calcScal(coord1, coord2)
}
\arguments{
\item{coord1}{set of coordinates1}

\item{coord2}{set of coordinates2}
}
\value{
scalefactor
}
\description{
Calculates the scale factor based on the coordinates of a reference image and the target image.
}
