% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calculateEvaluation}
\alias{calculateEvaluation}
\title{calculateEvaluation(objeto.seurat, mode = c("GTEM", "procrustes", "RVSSimageJ"),
listaCoordenadasNEW, listaCoordenadas,
patientType = c('unique','multiple'))}
\usage{
calculateEvaluation(
  objeto.seurat,
  mode = c("GTEM", "procrustes", "RVSSimageJ"),
  listaCoordenadasNEW = None,
  listaCoordenadas = None,
  patientType = c("unique", "multiple")
)
}
\arguments{
\item{listaCoordenadasNEW}{}

\item{listaCoordenadas}{}

\item{patientType}{}

\item{listaRawImages}{}

\item{listaTransImages}{}
}
\value{
Evaluation
}
\description{
Evaluate the alignment of images using various metrics (MSE, SSIM, etc.) for both raw and transformed images.
}
