% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedMethodsComparison.R
\name{calculateEvaluationPython}
\alias{calculateEvaluationPython}
\title{evaluation
calculateEvaluationPython(objeto.seurat, mode = c("STalign", "PASTE2"),
listaCoordenadasNEW, listaCoordenadas,
patientType = c('unique','multiple'))}
\usage{
calculateEvaluationPython(
  objeto.seurat,
  mode = c("STalign", "PASTE2"),
  listaCoordenadasNEW,
  listaCoordenadas,
  patientType = c("unique", "multiple")
)
}
\arguments{
\item{objeto.seurat}{Seurat object containing images and coordinates.}

\item{mode}{Evaluation mode: one of "GTEM", "procrustes", or "RVSSimageJ".}

\item{listaCoordenadasNEW}{List of new coordinates (optional, used mainly if mode != "RVSSimageJ").}

\item{listaCoordenadas}{List of original coordinates (optional, used mainly if mode != "RVSSimageJ").}

\item{patientType}{Patient type, affecting region size ("unique" or "multiple").}
}
\value{
Evaluation
}
\description{
Evaluate the alignment of images using various metrics (MSE, SSIM, etc.) for both raw and transformed images.
It uses the image of an object created by python (STalign or PASTE2).
}
