% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{calculateRVcoeff}
\alias{calculateRVcoeff}
\title{calculateRVcoeff()}
\usage{
calculateRVcoeff(
  modes = c("GTEM", "procrustes", "RVSSimageJ"),
  ims = c("2", "3", "4")
)
}
\arguments{
\item{modes}{A character vector indicating the deconvolution methods to be used. Default is c("GTEM", "procrustes", "RVSSimageJ").}

\item{ims}{A character vector indicating the image numbers to be processed. Default is c("2","3","4").}
}
\value{
A list of RV coefficients for each deconvolution method and image.
}
\description{
Calculates the RV coefficients for different deconvolution methods and images.
}
