% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{controlAlign}
\alias{controlAlign}
\title{controlAlign(image1)}
\usage{
controlAlign(image1)
}
\arguments{
\item{image1}{An image array (RGB or hex) used as reference.}
}
\value{
A list with the following control results:
\itemize{
\item \code{Positive control solution}: Metrics when image is compared to itself.
\item \code{Negative control solution}: Metrics when image is compared to a random negative control image.
\item \code{Movement control solution}: Metrics when image is compared to a shifted version.
}
}
\description{
Evaluate alignment metrics for a reference image with various controls.
}
