% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{createReferenceIfNeeded}
\alias{createReferenceIfNeeded}
\title{createReferenceIfNeeded(reference)}
\usage{
createReferenceIfNeeded(reference)
}
\arguments{
\item{reference}{object with the single-cell / single-nucelus reference for the deconvolution.}
}
\value{
reference object prepared for the deconvolution with RCTD
}
\description{
Checks if the reference for RCTD has been created correctly, and generates it if not.
}
