% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cropImage}
\alias{cropImage}
\title{cropImage(object, image_name, padding = 10)}
\usage{
cropImage(object, image_name, padding = 10)
}
\arguments{
\item{object}{A Seurat object containing spatial image data in the \code{@images} slot.}

\item{image_name}{A character string indicating the name of the image within the Seurat object to crop.}

\item{padding}{An integer specifying how many pixels of padding to add around the tissue region (default is 10).}
}
\value{
A modified Seurat object with the specified image cropped and updated.
}
\description{
Crops a square region from a spatial image within a Seurat object,
centered around the tissue coordinates, with optional padding. The cropped image and
corresponding coordinates are rescaled and updated in the object.
}
