% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cropImage_notscaled}
\alias{cropImage_notscaled}
\title{cropImage_notscaled(object, image_name, window_size)}
\usage{
cropImage_notscaled(object, image_name, window_size)
}
\arguments{
\item{object}{A Seurat object containing spatial image data in the \code{@images} slot.}

\item{image_name}{A character string indicating the name of the image within the Seurat object to crop.}

\item{window_size}{An integer specifying how many pixels of size will have the images.}
}
\value{
A modified Seurat object with the specified image cropped and updated.
}
\description{
Crops a square region from a spatial image within a Seurat object,
centered around the tissue coordinates. The cropped image and
corresponding coordinates are updated in the object, but not rescaled
so the real image size is the same in all.
}
