% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolutionRCTD_mergeFiles}
\alias{deconvolutionRCTD_mergeFiles}
\title{deconvolutionRCTD_mergeFiles()}
\usage{
deconvolutionRCTD_mergeFiles(
  modes = c("GTEM", "procrustes", "RVSSimageJ"),
  ims = c("2", "3", "4")
)
}
\arguments{
\item{modes}{A character vector indicating the deconvolution methods to be used. Default is c("GTEM", "procrustes", "RVSSimageJ").}

\item{ims}{A character vector indicating the image numbers to be processed. Default is c("2","3","4").}
}
\value{
A list of Seurat objects with deconvolution results added as assays.
}
\description{
Merges the deconvolution results from multiple files into a single list of Seurat objects.
}
