% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_example_data}
\alias{download_example_data}
\title{download_example_data(dest_dir = tempdir())}
\usage{
download_example_data(dest_dir = tempdir())
}
\arguments{
\item{dest_dir}{A character string indicating the destination directory where the
data should be saved and unzipped. Defaults to a temporary directory (\code{tempdir()}).}
}
\value{
A character string with the path to the directory where the data was extracted.
}
\description{
Downloads a ZIP file containing example data for the STIMA package
from the official GitHub repository and extracts it to the specified directory.
}
