% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedMethodsComparison.R
\name{escale}
\alias{escale}
\title{utils
escale(object.seurat, new_coords, patientType = c('unique','multiple'), N)}
\usage{
escale(object.seurat, new_coords, patientType = c("unique", "multiple"), N)
}
\arguments{
\item{object.seurat}{Seurat object containing the images.}

\item{new_coords}{Data frame containing the new coordinates to be rescaled.}

\item{patientType}{Type of patient data. Options are 'unique' or 'multiple'.}

\item{N}{Slice number to be rescaled.}
}
\value{
Data frame with rescaled coordinates.
}
\description{
This function rescales the coordinates of the images to match the original image dimensions after PASTE2 alignment.
}
