% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{EuclDist}
\alias{EuclDist}
\title{EuclDist(listaCoordenadas, nIm)}
\usage{
EuclDist(listaCoordenadas, nIm)
}
\arguments{
\item{listaCoordenadas}{list of coordinates}

\item{nIm}{list of image indices}
}
\value{
distance
}
\description{
Calculate Euclidean distance between two sets of coordinates.
}
