% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evalAlign}
\alias{evalAlign}
\title{evalAlign(image1, image2, listaCoordenadas, nIm)}
\usage{
evalAlign(image1, image2, listaCoordenadas, nIm)
}
\arguments{
\item{image1}{image1}

\item{image2}{image2}

\item{listaCoordenadas}{list of coordinates}

\item{nIm}{list of image indices}
}
\value{
solution
}
\description{
Evaluate alignment between two images using various metrics (MSE, SSIM, etc.).
}
