% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_cellTypesDeconvolution}
\alias{graph_cellTypesDeconvolution}
\title{graph_cellTypesDeconvolution(listaObjDeconv, cell.types, cell_colors)}
\usage{
graph_cellTypesDeconvolution(
  listaObjDeconv,
  cell.types = c("Astrcytes", "Microglia", "OPC", "Endothelial", "Neurons", "Pericytes",
    "Schwann", "Lymphocytes", "Oligodendrocytes", "Ependymal Cells", "Meninges"),
  cell_colors = c("#E41A1C", "#377EB8", "#4DAF4A", "#FF7F00", "#984EA3", "#FFFF33",
    "#A65628", "#F781BF", "#999999", "#66C2A5", "#FC8D62")
)
}
\arguments{
\item{listaObjDeconv}{A list of Seurat objects containing deconvolution results.}

\item{cell.types}{A vector of cell type names to be plotted.}

\item{cell_colors}{A vector of colors corresponding to the cell types.}
}
\value{
A list of plots for each cell type and image.
}
\description{
This function generates and saves spatial feature plots for cell type deconvolution results.
}
