% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_evalMetrics}
\alias{graph_evalMetrics}
\title{graph_evalMetrics(modes = c("GTEM", "procrustes", "RVSSimageJ","PASTE2","STalign"), patientType = c('unique','multiple'))}
\usage{
graph_evalMetrics(
  modes = c("GTEM", "procrustes", "RVSSimageJ", "PASTE2", "STalign"),
  patientType = c("unique", "multiple"),
  colors = c("green", "red", "lightblue", "orange", "pink", "violet", "purple",
    "mediumpurple", "purple4")
)
}
\arguments{
\item{modes}{A character vector indicating the deconvolution methods to be used. Default is c("GTEM","procrustes","RVSSimageJ","PASTE2","STalign").}

\item{patientType}{Type of patient data. Options are 'unique' or 'multiple'.}
}
\description{
graph_evalMetrics(modes = c("GTEM", "procrustes", "RVSSimageJ","PASTE2","STalign"), patientType = c('unique','multiple'))
}
