% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{hex2rgb_table_norm}
\alias{hex2rgb_table_norm}
\title{hex2rgb_table_norm(objetHex)}
\usage{
hex2rgb_table_norm(objetHex)
}
\arguments{
\item{objetHex}{}
}
\value{
rgba_array_norm
}
\description{
Convert hexadecimal color data into an RGB array with dimensions (x, y, 4), normalizes RGB values to a 0-1 range.
The function organizes color values (R, G, B, Alpha) into a 3D array.
}
