% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlipFunction.R
\name{inversionEjeX}
\alias{inversionEjeX}
\title{inversionEjeX(muestra1)}
\usage{
inversionEjeX(muestra1)
}
\arguments{
\item{muestra1}{A sample object containing the image and coordinates.}
}
\value{
A new sample object with updated coordinates and flipped image data.
}
\description{
This function flips the coordinates of a sample along the X-axis (newcol = 127 - col).
It also flips the image data along the X-axis.
}
