% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlipFunction.R
\name{inversionEjeXY}
\alias{inversionEjeXY}
\title{inversionEjeXY(muestra1)}
\usage{
inversionEjeXY(muestra1)
}
\arguments{
\item{muestra1}{A sample object containing the image and coordinates.}
}
\value{
A new sample object with updated coordinates and flipped image data.
}
\description{
This function flips the coordinates of a sample along both the X- and Y-axes
(newrow = 77 - row, newcol = 127 - col). It also flips the image data along both axes.
}
