% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlipFunction.R
\name{inversionEjeY}
\alias{inversionEjeY}
\title{inversionEjeY(muestra1)}
\usage{
inversionEjeY(muestra1)
}
\arguments{
\item{muestra1}{A sample object containing the image and coordinates.}
}
\value{
A new sample object with updated coordinates and flipped image data.
}
\description{
This function flips the coordinates of a sample along the Y-axis (newrow = 77 - row).
It also flips the image data along the Y-axis.
}
