% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mse}
\alias{mse}
\title{mse(image1, image2)}
\usage{
mse(image1, image2)
}
\arguments{
\item{image1}{image1}

\item{image2}{image2}
}
\value{
total_error
}
\description{
Calculate Mean Squared Error (MSE) between two images by comparing pixel values.
}
