% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mseGS}
\alias{mseGS}
\title{mseGS(image1, image2)}
\usage{
mseGS(image1, image2)
}
\arguments{
\item{image1}{image1}

\item{image2}{image2}
}
\value{
error
}
\description{
Compute the grayscale MSE between two images by first converting them to grayscale.
}
