% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedMethodsComparison.R
\name{PASTE2toSeurat}
\alias{PASTE2toSeurat}
\title{seurat2pythonCommunication
PASTE2toSeurat(object, patientType = c('unique','multiple'))}
\usage{
PASTE2toSeurat(object, patientType = c("unique", "multiple"))
}
\arguments{
\item{object}{Original object to be modified.}

\item{patientType}{Type of patient data. Options are 'unique' or 'multiple'.}
}
\description{
This function loads the PASTE2 data and creates a Seurat object.
}
