% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{resultProcrustes}
\alias{resultProcrustes}
\title{resultProcrustes(proc, mirrorx, mirrory, scale)}
\usage{
resultProcrustes(proc, mirrorx, mirrory, scale)
}
\arguments{
\item{proc}{Procrustes results}

\item{mirrorx}{TRUE or FALSE}

\item{mirrory}{TRUE or FALSE}

\item{scale}{TRUE or FALSE}
}
\value{
solucion
}
\description{
Get the results of Procrustes
}
