% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{rgb_table_norm2rgb_table}
\alias{rgb_table_norm2rgb_table}
\title{rgb_table_norm2rgb_table(objetRGB)}
\usage{
rgb_table_norm2rgb_table(objetRGB)
}
\arguments{
\item{objetRGB}{A numeric array of RGB values normalized between 0 and 1.}
}
\value{
An integer array with RGB values scaled to \link{0, 255}.
}
\description{
Generates the 0-255 RGB values for an image with RGB normalized values
}
