% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveImages}
\alias{saveImages}
\title{saveImages(object, out_dir = paste0(getwd(),"/original", format = c("tif","png"))}
\usage{
saveImages(
  object,
  out_dir = paste0(getwd(), "/original"),
  format = c("tif", "png"),
  name = NULL
)
}
\arguments{
\item{object}{A Seurat object containing spatial image data.}

\item{out_dir}{Character. Path to the output directory where the images will be saved.
Defaults to a folder named \code{"original"} in the current working directory.}

\item{format}{Save format TIFF or PNG.}

\item{name}{Image save name. If not included, the save name wll be the image name in the Seurat object.}
}
\value{
No return value. This function is used for its side effect: saving \code{.tif} or \code{.png} images to disk.
}
\description{
Extracts spatial images from a Seurat object (specifically from \code{object@images})
and saves them as \code{.tif} files in the specified output directory.
}
