% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedMethodsComparison.R
\name{saveSeurat_forAnnData}
\alias{saveSeurat_forAnnData}
\title{seurat2pythonCommunication
saveSeurat_forAnnData(seuratObj, patientType = c('unique','multiple'), samples = c("19","43","47","45"))}
\usage{
saveSeurat_forAnnData(
  object,
  patientType = c("unique", "multiple"),
  samples = c("19", "43", "47", "45")
)
}
\arguments{
\item{object}{Seurat object to be saved.}

\item{patientType}{Type of patient data to be saved. Options are 'unique' or 'multiple'.}

\item{samples}{Vector of sample identifiers. Default is c("19","43","47","45"). It is needed for multiple patients.}
}
\value{
None
}
\description{
This function saves the Seurat object in a format compatible with AnnData.
It saves the expression matrix, cell metadata, image coordinates, scale factors,
and gene metadata for each slice of the Seurat object.
}
