% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveSeurat_forAnnData_fromFolder}
\alias{saveSeurat_forAnnData_fromFolder}
\title{saveSeurat_forAnnData_fromFolder(RDSObjsFolder)}
\usage{
saveSeurat_forAnnData_fromFolder(RDSObjsFolder)
}
\arguments{
\item{RDSObjsFolder}{Folder containing the RDS Seurat objects.}
}
\value{
None
}
\description{
This function saves all the Seurat objects (RDS) from the given folder in a
format compatible with AnnData in a folder named 'tempfiles'.
It saves the expression matrix, cell metadata, image coordinates, scale factors,
and gene metadata for each slice of the Seurat object.
}
