% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectCoord}
\alias{selectCoord}
\title{selectCoord(image)}
\usage{
selectCoord(image)
}
\arguments{
\item{image}{}
}
\value{
list containing the coordinates of the selected points with two elements: x and y.
}
\description{
Allows the user to select multiple points on the provided image by clicking on it.
}
