% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedMethodsComparison.R
\name{selectCoordPython}
\alias{selectCoordPython}
\title{utils
selectCoordPython(image)}
\usage{
selectCoordPython(image)
}
\arguments{
\item{image}{}
}
\value{
list containing the coordinates of the selected points with two elements: x and y.
}
\description{
Allows the user to select multiple points on the provided image by clicking on it.
It uses the image of an object created by python (STalign or PASTE2).
}
