% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedMethodsComparison.R
\name{STaligntoSeurat}
\alias{STaligntoSeurat}
\title{seurat2pythonCommunication
STaligntoSeurat(object.STalign, object, patientType = c('unique','multiple'))}
\usage{
STaligntoSeurat(object.STalign, object, patientType = c("unique", "multiple"))
}
\arguments{
\item{object.STalign}{List object resulting from the STalign pipeline.}

\item{object}{Original object to be modified.}

\item{patientType}{Type of patient data. Options are 'unique' or 'multiple'.}
}
\description{
This function loads the STalign data and creates a Seurat object.
}
