\name{stima-package}
\Rdversion{1.1}
\alias{stima-package}
\docType{package}
\title{
Simultaneous Threshold Interaction Modeling Algorithm
}
\description{
This package enables you to estimate a regression trunk model. The core function is \code{stima}, which is also the name of the algorithm. The default model is a regression trunk model. A regression trunk model is an integration of a regression tree and a multiple regression model. Currently, the classification trunk model is being developed.  
}
\details{
\tabular{ll}{
Package: \tab stima\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-11-08\cr
License: \tab GPL-02\cr
LazyLoad: \tab yes\cr
}
The most important functions are \code{\link{stima},and \link{stima.control}}.
}
\author{
Elise Dusseldorp, Peter Neufeglise, and Claudio Conversano, with contributions of Kristof Meers and Cor Ninaber

Maintainer: elise.dusseldorp@tno.nl
}
\references{
Dusseldorp, E. Conversano, C., and Van Os, B.J. (2010). Combining an additive and tree-based regression model simultaneously: STIMA. \emph{Journal of Computational and Graphical Statistics, 19(3)}, 514-530.
}
\keyword{package}
\keyword{regression}
\keyword{tree}
\keyword{interaction}