% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedMethodsComparison.R
\name{alignmentSTalign}
\alias{alignmentSTalign}
\title{alignmentSTalign(object)}
\usage{
alignmentSTalign(object, patientType = c("unique", "multiple"))
}
\arguments{
\item{object}{first object with the slices/samples merged together
patientType Type of patient data. Options are 'unique' or 'multiple'.}
}
\description{
alignmentSTalign(object)
}
