% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{calcScal}
\alias{calcScal}
\title{calcScal(coord1, coord2)}
\usage{
calcScal(coord1, coord2)
}
\arguments{
\item{coord1}{A list with named components 'x' and 'y', representing the first set of coordinates.}

\item{coord2}{A list with named components 'x' and 'y', representing the second set of coordinates.}
}
\value{
A numeric value representing the optimal scale factor.
}
\description{
Calculate the scale factor between two sets of coordinates.
}
\details{
This function computes the optimal scale factor that minimizes the difference
between two sets of coordinates (coord1 and coord2) using a Brent optimization method.
}
