% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedMethodsComparison.R
\name{calculateEvaluationPython}
\alias{calculateEvaluationPython}
\title{evaluation
calculateEvaluationPython(objeto.seurat, mode = c("STalign", "PASTE2"),
listaCoordenadasNEW, listaCoordenadas,
patientType = c('unique','multiple'))}
\usage{
calculateEvaluationPython(
  objeto.seurat,
  mode = c("STalign", "PASTE2"),
  listaCoordenadasNEW,
  listaCoordenadas,
  patientType = c("unique", "multiple")
)
}
\arguments{
\item{listaCoordenadasNEW}{}

\item{listaCoordenadas}{}

\item{patientType}{}

\item{listaRawImages}{}

\item{listaTransImages}{}
}
\value{
Evaluation
}
\description{
Evaluate the alignment of images using various metrics (MSE, SSIM, etc.) for both raw and transformed images.
It uses the image of an object created by python (STalign or PASTE2).
}
