% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{controlAlign}
\alias{controlAlign}
\title{controlAlign(image1)}
\usage{
controlAlign(image1)
}
\arguments{
\item{image1}{image1}

\item{image2}{image2}

\item{listaCoordenadas}{list of coordinates}

\item{nIm}{list of image indices}
}
\value{
solution
}
\description{
Evaluate alignment metrics for a reference image with various controls.
}
