% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{createDeconvolutionLists}
\alias{createDeconvolutionLists}
\title{createDeconvolutionLists(object.seurat, split_by = "name", mode = c("GTEM", "procrustes", "RVSSimageJ"))}
\usage{
createDeconvolutionLists(
  object.seurat,
  split_by = "name",
  mode = c("GTEM", "procrustes", "RVSSimageJ")
)
}
\arguments{
\item{object.seurat}{A Seurat object after the alignment containing the data to be split.}

\item{split_by}{A character string indicating the metadata field to split the Seurat object by. Default is "name".}

\item{mode}{A character string indicating the deconvolution method to be used. Options are "GTEM", "procrustes", or "RVSSimageJ". Default is "GTEM".}
}
\value{
NULL (side effect: saves lists of Seurat objects to files)
}
\description{
Creates a list of Seurat objects for deconvolution analysis.
}
