% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluationComplete}
\alias{evaluationComplete}
\title{evaluationComplete(listaCoordenadasNEW, listaCoordenadas,
listaRawImages, listaTransImages,
patientType = c('unique','multiple'))}
\usage{
evaluationComplete(
  listaCoordenadasNEW,
  listaCoordenadas,
  listaRawImages,
  listaTransImages,
  patientType = c("unique", "multiple")
)
}
\arguments{
\item{listaCoordenadasNEW}{}

\item{listaCoordenadas}{}

\item{listaRawImages}{}

\item{listaTransImages}{}

\item{patientType}{changes the size of the region to be evaluated (feel free to change the values if they do not fit your data correctly)
(it was developed to fit with the example data)
unique: 200x200
multiple: 310x200}
}
\value{
Evaluation
}
\description{
Evaluate the alignment of images using various metrics (MSE, SSIM, etc.) for both raw and transformed images.
}
