% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{hex2rgb_table}
\alias{hex2rgb_table}
\title{hex2rgb_table(objetHex)}
\usage{
hex2rgb_table(objetHex)
}
\arguments{
\item{objetHex}{}
}
\value{
rgba_array
}
\description{
Convert hexadecimal color data into an RGB array with dimensions (x, y, 4).
The function organizes color values (R, G, B, Alpha) into a 3D array.
}
