% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{matrixComparison}
\alias{matrixComparison}
\title{matrixComparison(listaObjAnnot)}
\usage{
matrixComparison(listaObjAnnot)
}
\arguments{
\item{listaObjAnnot}{list of objects after deconvolution}
}
\value{
rv_long A data frame containing the RV coefficients and p-values for the comparison of cell types across different images.
}
\description{
Generates a matrix comparison of cell types across different images.
}
