% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mseGS}
\alias{mseGS}
\title{mseGS(image1, image2)}
\usage{
mseGS(image1, image2)
}
\arguments{
\item{image1}{A 3D array representing the first RGB image.}

\item{image2}{A 3D array representing the second RGB image.}
}
\value{
Numeric value representing the grayscale MSE between the two images.
}
\description{
Compute the grayscale MSE between two images by first converting them to grayscale.
}
