% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{rgb_table_norm2rgb_table}
\alias{rgb_table_norm2rgb_table}
\title{rgb_table_norm2rgb_table(objetRGB)}
\usage{
rgb_table_norm2rgb_table(objetRGB)
}
\arguments{
\item{objetRGB}{}
}
\value{
objetRGB
}
\description{
Generates the 0-255 RGB values for an image with RGB normalized values
}
