% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{solveCoord}
\alias{solveCoord}
\title{solveCoord(coord1, coord2, xmax, ymax)}
\usage{
solveCoord(coord1, coord2, xmax, ymax)
}
\arguments{
\item{coord1}{}

\item{coord2}{}

\item{xmax}{}

\item{ymax}{}
}
\value{
raw transformation parameters
}
\description{
Given two sets of coordinates (coord1 and coord2), and the maximum x and y values (xmax, ymax),
this function calculates cos(angle), sin(angle), dx, and dy by solving a system of equations.
}
