% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STIMA-package.R, R/alignment.R
\docType{"_PACKAGE"}
\name{STIMA}
\alias{STIMA}
\title{STIMA}
\usage{
STIMA(
  object,
  mode = c("GTEM", "procrustes", "RVSSimageJ"),
  scale = c(TRUE, FALSE)
)
}
\arguments{
\item{object}{first object with the slices/samples merged together}

\item{mode}{GTEM, procrustes, RVSSimageJ}

\item{scale}{TRUE or FALSE}
}
\value{
object.seurat aligned
}
\description{
Spatial Transcriptomics Image-based Methods for Alignment

Calculates the scale factor based on the coordinates of a reference image and the target image.
}
\details{
STIMA Package
}
\author{
Victor A. Gaya-Martín
maintainer Victor A. Gaya-Martín
}
